add modify
add modify

+++ SpringSoftware.Web/DAL/Manage/ImageManage.cs
@@ -119,8 +119,8 @@ namespace SpringSoftware.Web.DAL.Manage
 
         public static string GetOriginalImagePath(Picture picture)
         {
-            return  Path.Combine(AppDomain.CurrentDomain.BaseDirectory,"Images\\SaveUpload\\Product\\", picture.Id + Path.GetExtension(picture.FileName));
-          
+            return Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Images\\SaveUpload\\Product\\", picture.Id + Path.GetExtension(picture.FileName));
+
         }
 
         public static string GetOriginalImagePath(int pictureId)
@@ -132,10 +132,8 @@ namespace SpringSoftware.Web.DAL.Manage
         private static string GetThumbnailPath(Picture picture, int size)
         {
             return Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Images\\SaveUpload\\Product\\Thumbnails", picture.Id + "_" + size + Path.GetExtension(picture.FileName));
-
         }
 
-      
         public static string Get280PathByPictureId(int pictureId, IEnumerable<Picture> pictureList = null)
         {
             Picture picture = null;
@@ -177,6 +175,7 @@ namespace SpringSoftware.Web.DAL.Manage
         public static async Task<int> DeleteImage(int pictureId)
         {
             var picture = await _pictureDal.QueryByIdAsync(pictureId);
+            if (picture == null) return 0;
             var orginalImagePath = GetOriginalImagePath(picture);
             if (File.Exists(orginalImagePath))
                 File.Delete(orginalImagePath);
